backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
requests==2.31.0
beautifulsoup4==4.12.2
newspaper3k==0.2.8
aiohttp==3.9.1
PyPDF2==3.0.1
python-docx==1.1.0
faiss-cpu==1.7.4
numpy==1.24.3
openai==1.3.7
pydantic==2.5.0